+++ examples/resources/konnect_auth_server_clients/resource.tf
@@ -1,6 +1,6 @@
 resource "konnect_auth_server_clients" "my_authserverclients" {
   provider = konnect-beta
-  access_token_duration = 17749669
+  access_token_duration = 1458901
   allow_all_scopes      = true
   allow_scopes = [
     "247c0ad0-5487-46a8-bedf-a569c07c2442"
@@ -9,7 +9,7 @@ resource "konnect_auth_server_clients" "my_authserverclients" {
   grant_types = [
     "authorization_code"
   ]
-  id_token_duration = 11692687
+  id_token_duration = 961077
   labels = {
     key = "value"
   }